+++ src/main/java/net/geforcemods/securitycraft/renderers/ProjectorRenderer.java
@@ -7,7 +7,6 @@ import com.mojang.blaze3d.vertex.PoseStack;
 
 import net.geforcemods.securitycraft.ClientHandler;
 import net.geforcemods.securitycraft.blockentities.ProjectorBlockEntity;
-import net.geforcemods.securitycraft.blocks.ProjectorBlock;
 import net.geforcemods.securitycraft.renderers.state.ProjectorRenderState;
 import net.minecraft.client.renderer.SubmitNodeCollector;
 import net.minecraft.client.renderer.block.MovingBlockRenderState;
@@ -16,18 +15,12 @@ import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
 import net.minecraft.client.renderer.feature.ModelFeatureRenderer;
 import net.minecraft.client.renderer.state.CameraRenderState;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
 import net.minecraft.core.Vec3i;
-import net.minecraft.util.Mth;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
-import net.neoforged.neoforge.common.util.TriPredicate;
 
 public class ProjectorRenderer implements BlockEntityRenderer<ProjectorBlockEntity, ProjectorRenderState> {
-	public static final int RENDER_DISTANCE = 100;
-	private final TriPredicate<Integer, Boolean, Integer> yLoopBoundary = (projectionHeight, hanging, y) -> hanging ? y > -projectionHeight : y < projectionHeight;
-
 	public ProjectorRenderer(BlockEntityRendererProvider.Context ctx) {}
 
 	@Override
@@ -60,63 +53,26 @@ public class ProjectorRenderer implements BlockEntityRenderer<ProjectorBlockEnti
 
 		if (be.isActive() && !be.isEmpty()) {
 			Level level = be.getLevel();
-			BlockPos pos = be.getBlockPos();
-			boolean hanging = be.getBlockState().getValue(ProjectorBlock.HANGING);
-			Direction direction = be.getBlockState().getValue(ProjectorBlock.FACING);
-			int projectionHeight = be.getProjectionHeight();
-			int projectionRange = be.getProjectionRange();
-			int projectionOffset = be.getProjectionOffset();
-			Vec3i positionalOffset;
+			BlockPos projectorPos = be.getBlockPos();
+			AABB projectedBlocksArea = be.getProjectedBlocksArea().deflate(0.1D);
 
-			for (int x = 0; x < be.getProjectionWidth(); x++) {
-				for (int y = 0; yLoopBoundary.test(projectionHeight, hanging, y); y = hanging ? y - 1 : y + 1) {
-					if (!be.isHorizontal())
-						positionalOffset = getPositionalOffset(direction, x, y, projectionRange, projectionOffset);
-					else
-						positionalOffset = getPositionalOffset(direction, x, projectionRange - 16, y + 1, projectionOffset);
+			for (BlockPos relativePos : BlockPos.betweenClosed(projectedBlocksArea)) {
+				BlockPos projectionPos = projectorPos.offset(relativePos);
 
-					if (positionalOffset != null) {
-						BlockPos projectionPos = pos.offset(positionalOffset);
+				if (be.isOverridingBlocks() || level.isEmptyBlock(projectionPos)) {
+					MovingBlockRenderState movingBlockRenderState = new MovingBlockRenderState();
 
-						if (be.isOverridingBlocks() || level.isEmptyBlock(projectionPos)) {
-							MovingBlockRenderState movingBlockRenderState = new MovingBlockRenderState();
-
-							movingBlockRenderState.randomSeedPos = projectionPos;
-							movingBlockRenderState.blockPos = projectionPos;
-							movingBlockRenderState.blockState = be.getProjectedState();
-							movingBlockRenderState.biome = level.getBiome(projectionPos);
-							movingBlockRenderState.level = level;
-							renderPositions.add(new ProjectionInfo(positionalOffset, movingBlockRenderState));
-						}
-					}
+					movingBlockRenderState.randomSeedPos = projectionPos;
+					movingBlockRenderState.blockPos = projectionPos;
+					movingBlockRenderState.blockState = be.getProjectedState();
+					movingBlockRenderState.biome = level.getBiome(projectionPos);
+					movingBlockRenderState.level = level;
+					renderPositions.add(new ProjectionInfo(new BlockPos(relativePos), movingBlockRenderState));
 				}
 			}
 		}
 
 		state.renderPositions = renderPositions;
-	}
-
-	/**
-	 * Gets the offset of the block position to render a block at based on the projector's settings
-	 *
-	 * @param direction The direction the projector is facing
-	 * @param x The offset from the projectors position on the x-axis of the position at which to draw the fake block
-	 * @param y The offset from the projectors position on the y-axis of the position at which to draw the fake block
-	 * @param originalDistance The distance in blocks that the fake block is away from the projector (set by player)
-	 * @param originalOffset The offset in blocks that the fake block is moved to the side from the projector (set by player)
-	 * @return The offset of the fake block to be drawn, null if a non-horizontal direction was given
-	 */
-	private Vec3i getPositionalOffset(Direction direction, int x, int y, double originalDistance, double originalOffset) {
-		int distance = Mth.floor(originalDistance);
-		int offset = Mth.floor(originalOffset);
-
-		return switch (direction) {
-			case NORTH -> new Vec3i(x + offset, y, distance);
-			case SOUTH -> new Vec3i(x + offset, y, -distance);
-			case WEST -> new Vec3i(distance, y, x + offset);
-			case EAST -> new Vec3i(-distance, y, x + offset);
-			default -> null;
-		};
 	}
 
 	@Override
@@ -126,7 +82,9 @@ public class ProjectorRenderer implements BlockEntityRenderer<ProjectorBlockEnti
 
 	@Override
 	public AABB getRenderBoundingBox(ProjectorBlockEntity be) {
-		return new AABB(be.getBlockPos()).inflate(RENDER_DISTANCE);
+		AABB projectorBoundingBox = new AABB(be.getBlockPos());
+
+		return be.isEmpty() ? projectorBoundingBox.inflate(1.0D) : be.getProjectedBlocksArea().move(be.getBlockPos()).minmax(projectorBoundingBox).inflate(1.0D);
 	}
 
 	public record ProjectionInfo(Vec3i positionalOffset, MovingBlockRenderState movingBlockRenderState) {}